+++ src/main/java/kitchenpos/application/OrderService.java
@@ -1,9 +1,12 @@
 package kitchenpos.application;
 
 import kitchenpos.domain.*;
+import kitchenpos.dto.OrderLineItemRequest;
 import kitchenpos.dto.OrderRequest;
 import kitchenpos.dto.OrderResponse;
+import kitchenpos.exception.MenuNotFoundException;
 import kitchenpos.exception.OrderNotFoundException;
+import kitchenpos.exception.OrderTableNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,18 +14,41 @@ import java.util.List;
 
 @Service
 public class OrderService {
-    private final OrderFactory orderFactory;
     private final OrderRepository orderRepository;
+    private final MenuRepository menuRepository;
+    private final OrderTableRepository orderTableRepository;
 
-    public OrderService(final OrderFactory orderFactory, final OrderRepository orderRepository) {
-        this.orderFactory = orderFactory;
+    public OrderService(OrderRepository orderRepository, MenuRepository menuRepository, OrderTableRepository orderTableRepository) {
         this.orderRepository = orderRepository;
+        this.menuRepository = menuRepository;
+        this.orderTableRepository = orderTableRepository;
     }
 
     @Transactional
     public OrderResponse create(final OrderRequest orderRequest) {
-        Order order = orderFactory.create(orderRequest);
+        validateOrderLineItems(orderRequest);
+
+        final OrderTable orderTable = orderTableRepository.findById(orderRequest.getOrderTableId())
+                .orElseThrow(() -> new OrderTableNotFoundException(orderRequest.getOrderTableId()));
+
+        Order order = Order.create(orderTable);
+        addOrderLineItems(orderRequest.getOrderLineItems(), order);
+
         return OrderResponse.of(orderRepository.save(order));
+    }
+
+    private void addOrderLineItems(final List<OrderLineItemRequest> orderLineItems, final Order order) {
+        for (final OrderLineItemRequest orderLineItem : orderLineItems) {
+            Menu menu = menuRepository.findById(orderLineItem.getMenuId())
+                    .orElseThrow(() -> new MenuNotFoundException(orderLineItem.getMenuId()));
+            order.addOrderLineItem(menu, orderLineItem.getQuantity());
+        }
+    }
+
+    private void validateOrderLineItems(final OrderRequest orderRequest) {
+        if (orderRequest.isEmptyOrderLineItems()) {
+            throw new IllegalArgumentException("요청한 주문 항목이 없습니다");
+        }
     }
 
     @Transactional(readOnly = true)
