+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -2,9 +2,12 @@ package kitchenpos.application;
 
 import kitchenpos.domain.*;
 import kitchenpos.dto.OrderLineItemRequest;
+import kitchenpos.dto.OrderLineItemResponse;
 import kitchenpos.dto.OrderRequest;
 import kitchenpos.dto.OrderResponse;
+import kitchenpos.exception.MenuNotFoundException;
 import kitchenpos.exception.OrderNotFoundException;
+import kitchenpos.exception.OrderTableNotFoundException;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -27,9 +30,11 @@ import static org.mockito.BDDMockito.given;
 public class OrderServiceTest {
     public static final LocalDateTime TEST_CREATED_AT = LocalDateTime.of(2021, 12, 1, 12, 0);
     @Mock
-    private OrderFactory orderFactory;
-    @Mock
     private OrderRepository orderRepository;
+    @Mock
+    private MenuRepository menuRepository;
+    @Mock
+    private OrderTableRepository orderTableRepository;
     @InjectMocks
     private OrderService orderService;
 
@@ -42,21 +47,102 @@ public class OrderServiceTest {
             // given
             OrderTable orderTable = new OrderTable(1L, null, 4, false);
 
-            OrderLineItemRequest orderLineItemRequest1 = new OrderLineItemRequest(1L, 1);
-            OrderLineItemRequest orderLineItemRequest2 = new OrderLineItemRequest(2L, 1);
+            Menu 커플세트 = new Menu(1L, "커플세트", 16000, new MenuGroup(), Collections.emptyList());
+            Menu 혼밥세트 = new Menu(2L, "혼밥세트", 16000, new MenuGroup(), Collections.emptyList());
+            OrderLineItemRequest orderLineItemRequest1 = new OrderLineItemRequest(커플세트.getId(), 1);
+            OrderLineItemRequest orderLineItemRequest2 = new OrderLineItemRequest(혼밥세트.getId(), 1);
             List<OrderLineItemRequest> orderLineItemRequests = Arrays.asList(orderLineItemRequest1, orderLineItemRequest2);
             OrderRequest orderRequest = new OrderRequest(orderTable.getId(), orderLineItemRequests);
 
-            Order expectedOrder = new Order(orderTable, new ArrayList<>());
+            List<OrderLineItem> orderLineItems = new ArrayList<>();
+            Order expectedOrder = new Order(1L, orderTable, OrderStatus.COOKING, orderLineItems);
+            orderLineItems.add(new OrderLineItem(expectedOrder, 커플세트, orderLineItemRequest1.getQuantity()));
+            orderLineItems.add(new OrderLineItem(expectedOrder, 혼밥세트, orderLineItemRequest2.getQuantity()));
 
-            given(orderFactory.create(any(OrderRequest.class))).willReturn(expectedOrder);
+            given(orderTableRepository.findById(anyLong())).willReturn(Optional.of(orderTable));
+            given(menuRepository.findById(anyLong())).willReturn(Optional.of(커플세트), Optional.of(혼밥세트));
             given(orderRepository.save(any(Order.class))).willReturn(expectedOrder);
 
             // when
             OrderResponse order = orderService.create(orderRequest);
 
             // then
-            assertThat(order).isEqualTo(OrderResponse.of(expectedOrder));
+            assertThat(order.getOrderStatus()).isEqualTo(OrderStatus.COOKING);
+            assertThat(order.getOrderLineItems()).map(OrderLineItemResponse::getMenuId)
+                    .contains(커플세트.getId(), 혼밥세트.getId());
+        }
+
+        @DisplayName("주문 항목이 있어야 한다")
+        @Test
+        void requiredOrderItem() {
+            // given
+            OrderRequest orderRequest = new OrderRequest(1L, Collections.emptyList());
+
+            // when
+            ThrowableAssert.ThrowingCallable callable = () -> orderService.create(orderRequest);
+
+            // then
+            assertThatThrownBy(callable).isInstanceOf(IllegalArgumentException.class);
+        }
+
+        @DisplayName("주문 항목에 메뉴에 있는 메뉴만 있어야 한다")
+        @Test
+        void validateMenu() {
+            // given
+            OrderTable orderTable = new OrderTable(1L, null, 4, false);
+
+            OrderLineItemRequest orderLineItemRequest1 = new OrderLineItemRequest(1L, 1);
+            OrderLineItemRequest orderLineItemRequest2 = new OrderLineItemRequest(2L, 1);
+            List<OrderLineItemRequest> orderLineItemRequests = Arrays.asList(orderLineItemRequest1, orderLineItemRequest2);
+            OrderRequest orderRequest = new OrderRequest(orderTable.getId(), orderLineItemRequests);
+
+            given(orderTableRepository.findById(anyLong())).willReturn(Optional.of(orderTable));
+            given(menuRepository.findById(anyLong())).willReturn(Optional.empty());
+
+            // when
+            ThrowableAssert.ThrowingCallable callable = () -> orderService.create(orderRequest);
+
+            // then
+            assertThatThrownBy(callable).isInstanceOf(MenuNotFoundException.class);
+        }
+
+        @DisplayName("주문 테이블이 있어야 한다")
+        @Test
+        void validateTable() {
+            OrderTable orderTable = new OrderTable(1L, null, 4, false);
+
+            OrderLineItemRequest orderLineItemRequest1 = new OrderLineItemRequest(1L, 1);
+            OrderLineItemRequest orderLineItemRequest2 = new OrderLineItemRequest(2L, 1);
+            List<OrderLineItemRequest> orderLineItemRequests = Arrays.asList(orderLineItemRequest1, orderLineItemRequest2);
+            OrderRequest orderRequest = new OrderRequest(orderTable.getId(), orderLineItemRequests);
+
+            given(orderTableRepository.findById(anyLong())).willReturn(Optional.empty());
+
+            // when
+            ThrowableAssert.ThrowingCallable callable = () -> orderService.create(orderRequest);
+
+            // then
+            assertThatThrownBy(callable).isInstanceOf(OrderTableNotFoundException.class);
+        }
+
+        @DisplayName("주문 테이블이 비어있지 않아야 한다")
+        @Test
+        void notEmptyTable() {
+            // given
+            OrderTable orderTable = new OrderTable(1L, null, 4, true);
+
+            OrderLineItemRequest orderLineItemRequest1 = new OrderLineItemRequest(1L, 1);
+            OrderLineItemRequest orderLineItemRequest2 = new OrderLineItemRequest(2L, 1);
+            List<OrderLineItemRequest> orderLineItemRequests = Arrays.asList(orderLineItemRequest1, orderLineItemRequest2);
+            OrderRequest orderRequest = new OrderRequest(orderTable.getId(), orderLineItemRequests);
+
+            given(orderTableRepository.findById(anyLong())).willReturn(Optional.of(orderTable));
+
+            // when
+            ThrowableAssert.ThrowingCallable callable = () -> orderService.create(orderRequest);
+
+            // then
+            assertThatThrownBy(callable).isInstanceOf(IllegalArgumentException.class);
         }
     }
